Added map from dto->model. Fixed text concat bug
I was calling concat to append the text resources into the batch array,
however, I wasn't reassigning the result to the array variable.
Changed it so it used the spread operator in a push call instead

The mapping logic will allow objects to be properly instantiated with the
full use of class functions.